+++ belle-sip/src/belle_sip_internal.h
@@ -323,15 +323,15 @@ void belle_sip_socket_source_init(belle_sip_source_t *s,
                                   void *data,
                                   belle_sip_socket_t fd,
                                   unsigned int events,
-                                  unsigned int timeout_value_ms);
+                                  int64_t timeout_value_ms);
 void belle_sip_fd_source_init(belle_sip_source_t *s,
                               belle_sip_source_func_t func,
                               void *data,
                               belle_sip_fd_t fd,
                               unsigned int events,
-                              unsigned int timeout_value_ms);
+                              int64_t timeout_value_ms);
 belle_sip_source_t *belle_sip_fd_source_new(
-    belle_sip_source_func_t func, void *data, belle_sip_fd_t fd, unsigned int events, unsigned int timeout_value_ms);
+    belle_sip_source_func_t func, void *data, belle_sip_fd_t fd, unsigned int events, int64_t timeout_value_ms);
 void belle_sip_source_uninit(belle_sip_source_t *s);
 void belle_sip_source_reset(belle_sip_source_t *s);
 void belle_sip_source_set_notify(belle_sip_source_t *s, belle_sip_source_func_t func);
